Add memory tracking for StreamReader local buffers
We use some vectors as local buffers in the StreamReaders, but they were
not tracked in the system pool. This amount is inegligible, because
a new thread is assigned for each stage/task/split, and for each
thread, multiple StreamReaders would be created depending on the number
of columns and column type. The number of StreamReaders could be very
big. And each StreamReader could hold buffers of a few KB. Altogher
this can be a large amount.

This commit is to track this part of memory usage as part of the
PageSource system memory context.

We also try to include the instance sizes for the StreamReaders that
have at least one local buffer. The StreamReaders that don't have a
local buffer were not counted, i.e. all non-leaf level ones, like
ListStreamReader, MapStreamReader, and SliceDirectStreamReader, which
is a leaf level StreamReader but does not use a local buffer. This is
trying to follow the MemoryContext usage convention and not to confuse
people by updating the MemoryContext in each constructors. This missing
part could be a couple of dozens of MB on each worker node and not
significant.