Clear update count on getMoreResults method call
Original issue: https://youtrack.jetbrains.com/issue/DBE-6421

According to the JDBC documentation the getMoreResults method have to move
statement to the next "result" and the "result" in this case means not
only the ResultSet object. Also it affects the main condition which
determines that there is a result or not (getMoreResults() ||
getUpdateCount() != -1). This condition is also from JDBC specification.

So getMoreResults have to advance current results including update count.