Reduce rate of allocation in OperatorStats::add
OperatorStats::add() is a hot method, and it was previously calling
Duration::convertToMostSuccinctTimeUnit() and DataSize::succinctBytes()
resulting in non-trivial amount of memory allocation. The reason is
these methods iterate over TimeUnit.values() and Unit.values() enums,
and the values() call allocates a new array at each call. This change
fixes that by extracting these arrays as constants.

Here is the profiler output:

       bytes  percent  samples  top
  ----------  -------  -------  ---
111759400400   19.91%   304215  byte[]
 58400149752   10.41%   119837  java.util.concurrent.TimeUnit[]
 46858638328    8.35%   102423  io.airlift.units.DataSize$Unit[]
 45658652856    8.14%   120351  io.airlift.units.Duration
 39341134312    7.01%   157999  java.lang.Object[]
 19627675504    3.50%    71370  java.lang.String
 19524656360    3.48%    35741  io.airlift.units.DataSize