Refreshing file-based authZ
Adds an optional hook for the file-based authZ
that periodically refreshes the authZ and atomically
swaps out the existing authZ access control.

The initial implementation is added to the file-based
authZ mechanism but was implemented with the intention
that it can easily be added to any factory with a
few small code changes.

Using the refreshing authentication is as easy as
adding "security.refresh-enabled=true". This is currently
only supported for the file-based authorization.

Fixes #11687