Make ClusterMemoryManager resilient to memory accounting leaks
We have seen that if a finished query has non-zero memory reservation
on some worker (e.g., "leak") it will effectively disable the oom killer
as the lastKilledQuery will never be gone from the system
(isLastKilledQueryGone will return false). This change makes the
ClusterMemoryManager resilient to such memory leaks.

The change addds a ClusterMemoryLeakDetector and updates
ClusterMemoryManager to use the leak detector to determine whether the
lastKilledQuery has a leak, and if it has the ClusterMemoryManager
considers the lastKilledQuery as gone from the system, and continues
functioning as usual.