Fix memory accounting leak in AggregationOperator
If an AggregationOperator is used inside a TableWriterOperator
for stats collection the AggregationOperator will leak, as the driver
thread will only destroy the TableWriterOperator, and operators
rely on getting destroyed by the driver threads to free up their
allocations.

This change also closes the memory context of TableWriterOperator
in close(), however that's just for testing as the memory reserved
by that context will be released when the driver thread destroys the
TableWriterOperator operator context.