Add experimental config for disabling the reserved pool
This PR adds a new experimental config option for disabling the reserved
pool. The reason we have a reserved pool is to prevent deadlocks when
the general pool gets full. However, it's a source of inefficiency as a
good chunk of the heap just stays unused most of the time. Now that the
cluster oom killer works reasonably well (after making it resilient to
memory accounting leaks/bugs) we can run some experiments by disabling
the reserved pool -- when the general pool is full on some worker the
oom killer should just resolve the situation by killing a query. It's
possible that the killed query is retried by the client and cause oom
again, but the existing implementation is also already susceptible to
that issue, so this change doesn't make the situation any worse.