Remove min/max intersection heuristics from equality stats calculator
During TPCH/TPCDS benchmarking this optimization proved to performed worse that without it.
After investigation we found that cross-column correlation is not accounted for when applying
filtering for lo-max ranges (e.g: when filtering dimension table on date, we don't adjust row id
column range which is tightly correlated to date column). This caused great underestimation
of filtered rows.