Add spatial_partitions functions
spatial_partitions(KDBTree, Geometry) function returns a list of spatial
partition IDs for a given geometry. This function will be used to partition
inputs for a spatial join defined via ST_Contains or ST_Intersects.

spatial_partitions(KDBTree, Geometry, distance) function returns a list of
spatial partition IDs for a geometry representing a set of points within the
specified distance from the input geometry. This version will be used to
partition inputs for a distance query (e.g. a spatial join defined via
ST_Distance(a, b) <= r).

For a given geometry, a set of spatial partitions is defined as a set of
partitions whose extents intersect geometry's bounding box. For points, to
avoid unnecessary duplication of data when points fall on the borders of the
partition extents, spatial_partitions function returns at most one partition
by considering partition extents as rectangles open on the right and upper
sides.