Prevents sending large responses to client for certain optimized block types.
Resolves #11332
The change adds a method getUnoptimizedSizeInBytes to the Block interface for
getting the size of the content represented by the block as if no optimizations (such
as RLE or Dictionary encoding) were applied to it.
Using this provides a more accurate ballpark estimate of the data size that will
be sent to the client when Query::getNextResult is invoked.
For blocks other than RLE and Dictionary encoded the method returns getSizeInBytes().
More details at: https://github.com/prestodb/presto/issues/11332