Adds Block::getLogicalSize and limits size of response to client.
Resolves #11332
The change adds a method getLogicalSizeInBytes to the Block interface for
getting the size of a block's contents as if no optimizations (such as RLE or
Dictionary encoding) were applied to it.
This provides a more accurate estimate of the data size sent to the client when
Query::getNextResult is invoked.
For blocks other than RLE and Dictionary encoded the method returns getSizeInBytes().
More details at: https://github.com/prestodb/presto/issues/11332