Optimize geometry_union_agg and convex_hull_agg to reuse GeometryCursor
Current geometry aggregation functions will serialize the input OGCGeometry and perform simple union or convexHull on them iteratively.  This works, but it suffers from poor memory locality, and it has a nontrivial object allocation overhead.  The idea of this PR is to use the internal operators that are invoked inside these two methods, to reuse the operators, and save on object allocation and get better in-memory performance.

The most significant effect of reusing the cursors is that we have access to the optimizations that exist inside them, instead of throwing them away after every iteration of the operation.  For example, for union, with this changeset, the individual union operations will now be batched inside the union cursor, and once the internal batch size exceeds a constant value (10000), they will be unioned in aggregate.  This results in an increase in memory, but the performance wins should outweigh the relative cost.

The performance improvements are stark.  Here are the results for the old method:

```
Result "com.facebook.presto.plugin.geospatial.BenchmarkGeometryAggregation.benchmarkConvexHull":
  23.716 ±(99.9%) 0.572 ms/op [Average]
  (min, avg, max) = (22.057, 23.716, 25.605), stdev = 0.856
  CI (99.9%): [23.144, 24.288] (assumes normal distribution)
```

```
Result "com.facebook.presto.plugin.geospatial.BenchmarkGeometryAggregation.benchmarkUnion":
  628.715 ±(99.9%) 10.353 ms/op [Average]
  (min, avg, max) = (601.299, 628.715, 671.803), stdev = 15.495
  CI (99.9%): [618.362, 639.068] (assumes normal distribution)
```

Here are the results for the new method:

```
Result "com.facebook.presto.plugin.geospatial.BenchmarkGeometryAggregation.benchmarkConvexHull":
  4.556 ±(99.9%) 0.182 ms/op [Average]
  (min, avg, max) = (4.124, 4.556, 5.187), stdev = 0.273
  CI (99.9%): [4.374, 4.739] (assumes normal distribution)
```

```
Result "com.facebook.presto.plugin.geospatial.BenchmarkGeometryAggregation.benchmarkUnion":
  21.780 ±(99.9%) 1.320 ms/op [Average]
  (min, avg, max) = (20.106, 21.780, 31.502), stdev = 1.976
  CI (99.9%): [20.460, 23.101] (assumes normal distribution)
```