Limit the max line size in LineRecordReader
This is the fix for https://github.com/prestodb/presto/issues/11367

When reading large text files, in fairly rare cases Presto hive is
requesting allocation of byte[] of size 2,147,483,643 to 2,147,483,647.
This exceeds the VM limit which is Integer.MAX_VALUE-5. However apache
hadoop library uses Integer.MAX_VALUE as the limit and allows such
allocation. This results in the OutOfMemoryError but it's not caught
by Presto Hive. As a result of our JVM configuration, this results in
JVM restarts and kills all queries running on the cluster.

This commit is to
1) fix the above bug in Hadoop and allow loud failure when the line
is over the max limit;
2) make Presto depend on the snapshot version of the Hadoop library;
3) Set the default limit of text file line size 100MB in Presto.