Remove user cpu time tracking
Tracking user cpu time is expensive, as in Linux it is implemented in a way
of doing a full text parsing of the /proc/self/task/<process id>/stat.

The overall cpu time is implemented by reading lightweight pthreads cpu counter.

Traking user cpu time is 40x - 400x slower depending on platform and concurrency.
Removing tracking of the user time allows us to instrument operators, and have the
instrumentation by default at the cost of removing user cpu tracking. That seems to
be way more useful. Usefulness of having the user cpu time is very vague and arguable.