Plan FILTER clause projection as CASE expression
FILTER clauses in aggregates are today planned as mask field in aggregation.
This implies that we add a projection underneath the aggregate which will
project the aggregate expression and mask expression. However, this leads
to evaluating the aggregate expression assignment for all the input tuples,
whereas the assignment should be evaluated only for tuples where the filter
clause is true.