Revert "Improve Stripe Size Estimation for SliceDictionaryColumnWriter"
This reverts commit 9c0fe41281174b8bcd37378508e63228995befe0.

The reverted commit avoids overestimating dictionary index stream by
writing temporary dictionary index to a data stream and use the
buffered size of this stream as the estimation.

However, ORC spec requires dictionary to be sorted. Thus the dictionary
indices have to be re-indexed when flushing. In most case the estimated
size is close to the actual size. However under some rare corner cases,
the estimated size can be up to 6x smaller than the actual size, which
causes flushing huge stripes and becomes a reliability threat.

Here is an observed example: an column with ARRAY(VARCHAR), where
each cell contains the same array with about 30,000 elements.
The array is unsorted. Thus the temporary dictionary index
is [1, 2, 3, ...], while after re-indexing, the dictionary index
is a random shuffle.  The dictionary index streams exhibit
significantly different compression ratios under ZLIB.

The conclusion is with the re-indexing as the last step, it is
extremely difficult, if not impossible, to accurately estimate
the index stream size after compression.

We decide to fall back to the previous mechanism which tends to
over-estimate the stripe size, for the following reasons:

 1. While the previous heuristics flush small compressed stripes,
the uncompressed stripe size is as expected thus each stripe still
has enough data for computation engine to work on.

 2. Writing arbitrarily huge stripe is worse to Presto engine from
reliability perspective.

Note DWRF doesn't require dictionary to be sorted, in the future we can
add an option to write unsorted dictionary when writing DWRF files.