Add SphericalGeography type
The SphericalGeography type provides native support for spatial features
represented on "geographic" coordinates (sometimes called "geodetic"
coordinates, or "lat/lon", or "lon/lat"). Geographic coordinates are spherical
coordinates expressed in angular units (degrees).

The basis for the Geometry type is a plane. The shortest path between two
points on the plane is a straight line. That means calculations on geometries
(areas, distances, lengths, intersections, etc) can be calculated using
cartesian mathematics and straight line vectors.

The basis for the SphericalGeography type is a sphere. The shortest path
between two points on the sphere is a great circle arc. That means that
calculations on geographies (areas, distances, lengths, intersections, etc)
must be calculated on the sphere, using more complicated mathematics. More
accurate measurements that take the actual spheroidal shape of the world into
account are not supported.

Add a variant of ST_Distance function that takes two spherical points and
returns great-circle distance in meters.

Note: spatial joins do not apply to spatial functions operating on
SphericalGeography yet. Support will be added later.

Resolves: Issue #11580