Do not return wrong stats/cost for partial aggregations
The stats and cost calculation formulas for `AggregationNode` are
correct for SINGLE step aggregation only. This commit causes stats
calculation to (more correctly) return unknown in all other cases.

Technically, cost calculation for FINAL aggregation was correct, even
though the stats calculation was not. For this reason, cost calculation
was retained, even though it is currently ineffective.