Reimplement JoinBridgeManager to fix bugs and simplify logic
In this commit, JoinBridgeManager is reimplemented.

The commit fixes execution failure of build-outer join when probe side
is grouped execution but build side is not.

The commits also simplifies the logic.
JoinBridgeManager had two parts: DataManager and LifecycleManager.
The responsibility between the two was never clearly separated, leading
to unnecessary complexity and confusion.

Some tests violate the assumption that operator/driver instantiation
happens after all OperatorFactory are created. Prior to this change,
those tests goes around lifecycle management of the join operator.
But that's no longer possible now that the lifecycle management is built
into the JoinBridgeManager. As a result, changes were made to such tests.