Add method getLogicalSize to Block
The change adds a method getLogicalSizeInBytes to the Block interface
for getting the size of a block's contents regardless of the internal
representation. For Dictionary and RunLengthEncoded blocks this can vary
substantially from the value returned by getSizeInBytes.

A subsequent commit will use this method to resolve #11332 as this will
provide a better estimate of data size sent to client in
Query.getNextResult()