Avoid making wild guesses when estimating filters and joins
If the stats are missing, or it is simply not possible to give any reasonable
estimate when all of the conjuncts are unsupported, return uknown estimate
instead of 0.9 filter factor.

Preserve 0.9 filter factor for every additional conjunct when at least one
conjunct can be reasonably estimated.

This change is supposed to protect from making wild guesses when estimating
JOINs. Since the JOIN cardinality is estimated as cross_product * filter_factor.
Making wild guesses can result in completely unpredictable JOIN order, as one
of the joins can be greatly overestimated.