Count splits for unstarted source nodes as queuedDrivers
When a split for a partitioned pipeline is delivered to a worker,
conceptually, the worker would have an additional driver.
The queuedDrivers field in PipelineStatus is supposed to represent this.

At some point, this was literally how SqlTaskExecution works.
However, the internal implementation of SqlTaskExecution has changed.
The SqlTaskExecution may not immediately instantiate a driver for the split.
But PipelineContext wasn't updated. When it reports PipelineStatus,
it still computes the number of drivers that haven't started execution
as "queuedDrivers" (and "queuedPartitionedDrivers"). As a result,
some split delivered to workers are not reflected in the queuedDriver number
because they haven't been instantiated into a driver yet.

Fixing the queuedDrivers field in PipelineStatus has the following
consequences:

* The commit affects client stats
* The commit restores split scheduling back pressure when there is
  more than one partitioned source in a Stage (e.g. colocated join).
  NodeScheduler.selectDistributionNodes depends on
  NodeAssignmentStats.getTotalSplitCount, which is effectively queuedDrivers +
  recent assignment. While recent assignment would increment, it is reset
  to zero whenever the split is delivered to workers. Prior to this commit,
  due to the bug described above, queuedDrivers would
  be zero for all partitioned sources other than the first one
  (in source scheduling order) until the first one finishes scheduling.