Introduce BucketNodeMap
Splits from the same buckets are scheduled to the same node to allow
bucketed execution. Previously, NodeSelector#computeAssignments takes
NodePartitionMap, which contains extra information.

This commit introduces BucketNodeMap to represent the mapping in a
separate class. This also opens opportunities for engine to schedule
bucket execution in a more flexible way.