Move MapBlock related classes to a new package
This is the preparation commit for the lazy MapBlock hashtable project.
It creates a new package com.facebook.presto.spi.block.Map and move all
map related classes to this package.

For simple read/write queries involve maps, we don't need to build the
hashtable. The idea is to build hashtable on the fly when the seekKey
methods in SingleMapBlock are called, and MapBlockBuilder need to
rebuild the missing hashtable when closeEntry() is called with a
passed in hashtable. However synchronizing the seekKey() and
closeEntry() methods is very complex and error prone. To reduce the
complexity of synchronization, we will make MapBlockBuilder build the
hashtable all the time and only synchronize when multiple threads are
trying to create the hashtable when calling seekKey(). In a seperate
project, we will avoid MapBlockBuilder to build the hashtables in the
PartitionOutputOperator. To make this way work, we need to make
SingleMapBlock to be able to see the hashtable inside of the MapBlock.
Instead of making it public, we will create a new Map package and make
the hashtable visible to this package.