Mark lifespan as completed after output is acknowledged
Currently, a lifepsan reported to be completed might still have
pending data in OutputBuffer. This makes it difficult to decide
lifespans to be retried in the scenario of failure recovery.

This is because a lifespan that fails after all drivers finish
should still be retried if its output has not been acknowledged
by the receiver.