Check yield signal every 5s in generated project method
Checking the yield signal for every row results in a performance hit.
This change reduces this performance hit by doing the check every 5s.

tpch q1, q10, and q15 (arbitrarily picked) with a scale factor of 100
are used to assess the performance of this change. Here are the
improvements/savings (none of the tests resulted in worse performance)
in aggregate CPU time spent in the ScanFilterAndProject operators in
these queries. When there are multiple numbers for a query it means
that there are multiple stages with ScanFilterAndProject operators
(these numbers are aggregated stage-wide).

q1: 4%
q11: 20%, 17%, 31%, 19%
q15: 19%, 1%, 1%

Finally, here's the generated project method with this change:
https://gist.github.com/nezihyigitbasi/b015a718e49280ed2079e14b2946e1a1