Disable stats and cost calculators by default
Stats calculator fails on assertions for complex queries, thus it is not production ready yet.

The https://github.com/prestodb/presto/pull/11511 changes the place where the stats are computed to be
displayed in the final plan. Before this patch, the stats were computed in QueryMonitor, when printing a final plan.
Before if the stats couldn't be computed for whatever reason, only the text plan generation would fail.

Currently, when the stats calculator is invoked for every query during the initial planning, queries may
be failing, even when the CBO is not used.

This change disables stats calculator by default. It can be enabled back with a session property on per query basis.