Make JsonFieldDecoder extensible again.
At some point, the members of JsonValueProvider were changed from
protected to private. This broke any code that actually builds on top of
this class because the value inside the class was no longer accessing.

It also breaks out the clamping logic that ensures that values that map
on non-long columns stay within the value range into its own class;
there are legitimate use cases to extend JsonValueProvider that do not
need to be burdened with a c'tor that requires values for a very
specific use case.