Update the batch resizing heuristic in PageProcessor
The yield check in the generated projection tight loop may have negative
impact on the performance. The yield check was introduced for expensive
projections so that the engine can yield those. However, that check is
not really necessary for cheap projections and having a branch in the
tight for loop seems to hurt the performance.

This change removes the yield check from the generated projection loop,
and to handle the expensive expressions properly it adds the cost of the
evaluated expression as another variable to the batch resizing heuristic
in PageProcessor. With that for expensive expressions Presto now uses
small batch sizes, and uses larger batch sizes for cheaper expressions.

Removing the yield check from the projection loop improves the performance
as much as ~5% with microbenchmarks and various production queries.

Benchmark                                   (expressionProfilerEnabled)  Mode  Cnt   Score   Error  Units
ExpressionProfilerBenchmark.pageProjection                        false  avgt   15  37.672 ± 4.401  us/op
ExpressionProfilerBenchmark.pageProjection                         true  avgt   15  35.336 ± 4.724  us/op