Improve the performance of generated projection loop
The yield check in the generated projection tight loop may have negative
impact on the performance. The yield check was introduced for expensive
projections so that the engine can yield those. However, that check is
not really necessary for cheap projections and having a branch in the
tight for loop seems to hurt the performance.

This change introduces an ExpressionProfiler to determine whether a
projection is cheap/expensive, and as a result, in the best case where
the projection is cheap the generated projection code can execute a
simple tight loop for expression evaluation, and get rid of the yield
check entirely.

This change has been benchmarked with various production queries and
microbenchmarks and it results in up to ~5% improvement in performance.

Benchmark                                   (expressionProfilerEnabled)  Mode  Cnt      Score      Error  Units
ExpressionProfilerBenchmark.pageProjection                        false  avgt   15  25176.150 ± 2187.975  ns/op
ExpressionProfilerBenchmark.pageProjection                         true  avgt   15  23767.123 ± 1480.502  ns/op