Fix memory leak for failed queries
Queries that were canceled via SqlQueryExecution#cancelQuery or failed via SqlQueryExecution#fail
are not getting evicted in the QueryTracker. The queries are not getting evicted as they never
being placed into the eviction queue, as the fianl query info callback is never called.

Implicitly setting the query info by calling getQueryInfo in the abandoned queries handler mitigates the issue,
as eventually all stage info's arrive as finals, and the final query info is set.