Ignore failures in stats and cost calculator
Currently stats and costs are pre-computed for every query.
There are still some bugs in stats calculator that are being
triggered in very rare cases that are incredibly hard to catch.
Given that the stats calculator has to be run for every query, it
is better to be fail safe. As a failure in stats calculation must
not result in failing query.