FIx deployment of presto-jdbc artifact
The presto-jdbc artifact shades the resulting jar and creates a new
source jar. There is a bug in maven (MSHADE-195 and MSHADE-259) that
causes this source artifact to be deployed twice. This is not a problem
for SNAPSHOT or releases to a staging repo but it will fail for releases
to a regular repo where a non-SNAPSHOT version can not be overwritten.

This fixes the deployment to only deploy the source repository
once. With this fix, a deployment e.g. to an internal repo succeeds.