Accelerate build-side geometries in spatial join
Build gemetries are often used repeated for spatial relation checks in a
spatial join.  Accelerating a geometry is a one-time cost which reduces
the complexity of these checks from `O(n)` to `O(lg n)` or even `O(1)`,
where n is the number of vertices.  When a geometry is accessed for the
join, accelerate it (an idempotent operation).

Upgraded ESRI library to 2.2.2 (from 2.2.1) to pick up a fix in
acceleration logic.