Fix assertion in tryUpdateUserMemory method
When DefaultQueryContext::tryUpdateUserMemory() is called with a delta
of 0 and when the general pool is full, the updateUserMemory call will
return a future that's not done. However, tryUpdateUserMemory was simply
assuming that for that case the future must be done and asserting on
that, which causes query failures. This change fixes that.

It makes sense for updateUserMemory to return a future that's not done
when we reserve 0 bytes and the pool is full, so that the caller knows
that the pool is full and waits on that future to make the next
reservation. Therefore, for the delta == 0 case it's not correct to
simply assume that future returned from updateUserMemory is done.