Remove hashCode and equals from AccessControlInfo
AccessControlInfo does not look as value object and hence it should not
implement hashCode and equals methods.

Also it is used only in
Map<AccessControlInfo, Map<QualifiedObjectName, Set<String>>>.
It looks like  a bug, because if we would get two equal
AccessControlInfos, then we could lose one
Map<QualifiedObjectName, Set<String>> and so we would skip one
access check.