Remove unsupported types on write path in JDBC connectors
These types are not supported in
`com.facebook.presto.plugin.jdbc.JdbcPageSink#appendColumn`, so cannot
be written.

Since JDBC based connectors do not support CREATE TABLE except for
CREATE TABLE AS, it makes no sense to attempt to create columns we
cannot write to.

Note: attempt to support these types must be well test-covered on
per-database basis, so they should not be supported in a generic way
anyway.