Upgrading parquet version to 1.10.0
Presto is on an old version of Parquet, so it is missing out on some of
the newer features, such as using column ids instead of column names,
which are required for developing new connectors like iceberg. Parquet
moved to Apache so all the classes have a different namespace now, so
we can keep both the old and new Parquet classes on the classpath.
This allows us to upgrade the Parquet version for native Parquet read
path and let the Hive read and write path continue using the old Parquet
version until we upgrade the Hive version.