Assume Domain from file cannot be isNone()
This commit changes code behavior in the unlikely edge case where:

1. Domain constructed from file stats is empty (`isNone()`)
2. we did not skip the file/stripe based on `numberOfRows == 0`
   condition.
3. there is no effective predicate for a column

Previously, the code would construct realize the `Domain` constructed
from file stats `isNone()`. Now, the `Domain` construction will not be
attempted, as the code short-circuits when there no effective predicate
for a column. This change is based on assumption that the `Domain`
constructed from file stats cannot `isNone()` in any practical
situation.