Use known Presto Type instead of reconstructing
When creating `Domain` from file stats, we need to know actual Presto
Type of the column, as Types must match for `Domain#intersect`.

Instead of reconstructing the type based on the type in the file, which
is not always possible, as Parquet does not distinguish between INTEGER,
SMALLINT etc, use the Type passed with effective predicate.