(1/3) eclipse/m2e compatibility: trivial changes
1. RecordingHiveMetastore.java had two places where it's doing

    CacheBuilder.<K, V>newBuilder().build()

even though `newBuilder()` isn't generic. I was surprised to find that
JLS explicitly allows this (JLS 11th ed, sec 15.12.2.1), but ecj issues a
warning. Move the `<K, V>` to `build()` which actually is generic.

2. In TestArrayOperators and TestMapOperators, there are two places
where ecj is doing the inference differently from javac. I *suspect* ecj
is wrong here because, while I'm not curious enough to dredge through
JLS chapter 18, the fact that it reproduces in
https://gist.github.com/jesboat/90bda38a0f2d2208a8b114f80943ad34
only if `somehowRequired()` is present means at least something wonky is
going on.