Repartitioning improvement
Before this change, the PartitionedOutputOperator write rows of each
partition to BlockBuilders first, then serialize them into SerializedPage
before appending them to OutputBuffer. This change will skip the
BlockBuilder buliding step and directly append values to the buffers.
We calculate the row sizes for each incoming page, and only write the rows
that would fit in the buffers. When the buffers are full, we flush them
to SerializedPage and send it out.

This commit introduces several new methods in Block interface because we
wanted to do batch processing in tight loops. There were existing study
showing gains working on raw arrays in "batch" mode in tight loop over
working on Blocks in "row" mode. We will do additional JMH benchmarks
to compare the effect for repartitioning case.

Please also note that this commit does not contain the buffer pool logic yet.
The buffers are with fixed size and is assumed to have enough space for
now.