Remove CURRENT_ROLE from the reserved keywords
The SQL standard allows queries like:

```
postgres=# SELECT CURRENT_ROLE;
 current_role
--------------
 postgres
(1 row)
```

That's why the `CURRENT_ROLE` is a reserved keyword in the SQL standard.

Currently Presto supports the `CURRENT_ROLE` keyword only in
`CREATE ROLE`, `GRANT` and `REVOKE` statements. So it is possible to
parse the `CURRENT_ROLE` token as a keyword in those statements, while
parsing the same token as an identifier in any other queries.

However this is only a temporary solution, as at some point the scope
of the `CURRENT_ROLE` might be extended to the SELECT queries as well.