Reduce the identifier length in generated code for nested columns
Currently the type signature is used as name which could be long if the
type is a nested type. We recently hit the limit of identifier length in
ASM [1] on one of the wide tables which has nested column deferences in
project/filter.

Change the name to a short name based on the Java object type of the `Type`,
as the instruction definition is based on the Java object type. This reduces
the size of the code generated too.

Bytecode for the test query used in unittests:
Before fix: https://gist.github.com/vkorukanti/64bfab1842c192fdb21f12f54006def3#file-before_fix
After fix: https://gist.github.com/vkorukanti/64bfab1842c192fdb21f12f54006def3#file-after-fix

[1] https://gitlab.ow2.org/asm/asm/blob/master/asm/src/main/java/org/objectweb/asm/ByteVector.java#L246

address review comments:
1. Use the existing logic up to some size limit (maybe 20 bytes), and after that, `use type.getTypeSignature().getBase()`
2. Increase the width of the nested column to repro the issue in test
3. Revert change to SqlTypeByteCodeExpression.formatOneLine