Replace ValuesNode::Expression with RowExpression
Original Expression is a special RowExpression implementation holding an
Expression object that cannot be translated directly while constructing
PlanNode. Typical examples are cases with `WITH` clauses or alias that
the type or arguments information are not directly available given the
substrees are not formed yet. OriginalExpression should not exist after
optimization or serialized over the wire. All OriginalExpressions should
be translated to other corresponding RowExpression objects ultimately.

Wrap Expression with OriginalExpression in ValuesNode. Place them with the
actually RowExpressions once we have reached the end of optimization.