Enforce scan batch by exception on exceeding space budget
OrcRecordReader.getNextPage tracks the average bytes per rows of
initial qualifying set. The batch size is adapted to stay within
expected soft limits. If a hard limit is exceeded an exception is
thrown and the streams are reset to the beginning of the current row
group. This behavior only works when the streams are checkpointed,
hence ValueStreamSources are not used even if the stripe consisted of
a single row group.

This replaces the StreamReader result truncation mechanism.