Encode PlanFragmentId as string in JSON
The "Create stage id based on the plan fragment id" commit changed the
way the PlanFragmentId is encoded in JSON:

https://github.com/prestodb/presto/commit/326f80539e42847dd9292d7847da4df8a331a55e

Before this patch the value had been encoded as a String, and after
this patch the PlanFragmentId started to be encoded as a number.

This breaks the "Live Plan" UI, as the "Live Plan" controller doesn't
match the remote source ids that are still encoded as strings
(see JsonRenderedNode):

https://github.com/prestodb/presto/blob/master/presto-main/src/main/java/com/facebook/presto/sql/planner/planPrinter/JsonRenderer.java#L53
https://github.com/prestodb/presto/blob/master/presto-main/src/main/java/com/facebook/presto/sql/planner/planPrinter/JsonRenderer.java#L64
https://github.com/prestodb/presto/blob/master/presto-main/src/main/resources/webapp/src/components/LivePlan.jsx#L238

An alternative solution is to encode JsonRenderedNode#remoteSources
as a number as well. However, since all other ids (PlanNodeId, StageId)
are encoded as a string, it seems to be more convenient to also have
the PlanFragmentId encoded as a string.