Fix skipping and values sizing in diverse scalar stream readers
Resets the toSkip variable after skipping to the next row in the input
qualifying set. The error does not manifest if all rows are selected.

Sets the expected output size according to the input qualifying set.

Adds a test. Note that the test is not supported by H2 and requires comparing Presto with and without Aria scan.