Rename finish() to noMoreSplits() for PerBucket
Having both finish() and isFinished() is confusing because they are
talking about different events. finish() is called when
BackgroundHiveSplitLoader notifies HiveSplitSource that no more
splits will be enumerated, whereas isFinished() is used to determine
if all the splits have been fully fetched. We renmae finish() to
noMoreSplits() to avoid this confusion.