Introduce InternalPlanNode and its visitor
PlanNode will be moved to SPI. InternalPlanNode inherits PlanNode to
represent the nodes that cannot be exposed to connectors. As a first
step, FilterNode and TableScanNode inherit PlanNode directly to indicate
they will be moved to SPI. More detailed javadoc is added to these two
nodes; Guava dependency is removed from them.