Avoid exception message duplication in PageResponseHandler
When a PageTransportErrorException is thrown, today it will be wrapped
in another PageTransportErrorException with the original exception
message copied.

This is unnecessary since the original message is contained in cause.
When the exception message is long (e.g. it may contain stacktrace from
the remote host), useful information can be omitted.