Introduce SplitSchedulingContext in getSplits() SPI
Replace SplitSchedulingStrategy with SplitSchedulingContext, which
contains both the SplitSchedulingStrategy and a boolean of
'schedulerUsesHostAddresses' that tells the connector whether the
scheduler will take advantage of host addresses for remotely-accessible
splits.  If it won't the connector can return an empty list of host
addresses without any performance cost. This is valuable because
Block address information can take up a lot of memory in a split, which
is especially problematic for grouped execution which requires
generating all splits at once.