Fix race condition in the SqlQueryScheduler
The `scheduling` is supposed to ensure that only a single
scheduling happens at any moment of a time.

If the finally block is executed after the next streaming section
has already started scheduling, the "scheduling" flag will be set
to "false" despite the scheduling of the second streaming section
is still in process.