Fix time format string in TestDateTimeFunctionsBase
`%M` is the code for the current minute; `%m` is the month. `%Y-%M` is
thus essentially never a format string you want.

This isn't actually a bug since this is testing that the code errors,
and won't actually examine the contents of the format, but I found it
in a large-scale grep for this common bug, and fixing it resolves the
slim chance of someone cargo-culting it, or doing the same grep and
having to do the same work.