Change jar version to work with branch based release model
The previous model involved using git-describe which broke as we
moved to branch based releases since all of our tags are now on the
release branches rather than master.

Even before that, the versions generated used the previous version
number of the release rather than the current maven version.

Changed that to now use a combination of maven version and commit id.