Enable tuple domain extraction for subfields
The Hive connector is supposed to be able to extract range filters
(tuple domains) that apply to subfields, e.g.

    SELECT a from t where b[1] = 10

This subfield extraction was failing in some cases because expressions
were not optimized and could have an unpredictable shape and didn't
resolve to a constant expression when they could have.

This commit applies RowExpressionOptimizer to these expressions
so they can be recognized and extracted.

Fixes: https://github.com/prestodb/presto/issues/12917
PR: https://github.com/prestodb/presto/pull/12936