Fix mis-optimization in predicate push down
When an expression includes variables from both sides of a join,
we could push down predicate incorrectly. For example, for
  IF (v1 = 1, v2 > 1) AND v1 = 1
if v1 is from t1, and v2 is from t2, since we know v1 = 1, we could
rewrite the expression to
  IF (1 = 1, v2 > 1) AND v1 = 1
This would result in the expression IF (1 = 1, v1 > 1) be pushed down
to t1, which would always evalute v2 > 1, rather than only evaluate
it when v1 = 1 is true.

We fix this by checking variableScope before doing equality inference
expression rewrite, rather than after. This might cause regression in
cases that such push down is safe, but it would ensure correct results.