Fixed Histogram Data Structures
In certain cases, it's beneficial to find the histogram of values
within a known range. In these cases, a fixed-range histogram, where
the minimum and maximum are static, is very efficient. This diff adds
two such histograms, for double values:
1. A histogram storing, for each bucket, the total sum of weights
   falling into the bin
2. A histogram storing, for each bucket and weight, the total number
   of times the combination was seen

These two data structures are useful for calculating statistics on a
histogram. The second, in particular, allows performing Jacknife bias
correction. A further diff on differential entropy will use these data
structures.