Fetch SearchHits scroll by scroll as needed than all at once in Elasticsearch
Currently we fetch all SearchHits in one call and keep them in memory. This causes
too much delay in getting first record and memory pressure. Write a custom iterator
that fetches the SearchHits as needed. One downside with this approach is the
scroll could get timedout, but there is a way to config custom scroll time out in
plugin config.