Create zero-row file in parallel
Buckets of a bucketed table in Hive were initialized as zero-row files serially,
which is very slow.

Now the empty file creation logic in HiveMetadata is factored out as
HiveZeroRowFileCreator, which creates files in parallel in a configurable
way similar to HiveStagingFileCommitter. By default the executor has
maximum 20 threads to create files, and can be configured by variable
hive.max-concurrent-zero-row-file-creations. Considering a 50ms file creation
latency, this feature significantly improves the performance.