Fix NPE in spatial joins with GeometryCollections
The code which gets a (JTS) Envelope from a GeometryCollection
raised an NPE: getting an Esri Envelope from an OgcGeometry that
wraps a GeometryCollection returns null.  We correct for this
in the GeometryUtils.getEnvelope; this commit delegates getting
the Esri envelope to that function, and converts it to a JTS
Envelope in GeometryUtils.  It also adds some tests, including the
breaking GeometryCollection (a reduced test case that caused the
bug in production).