Throw exception when creating an empty spatial partition
If the spatial partition query yielded 0 rows, an empty spatial
partition would be made.  It's unclear what an empty spatial partition
is!  Correspondingly, it was serialized as null, and then caused errors
during deserialization and use.

This throws an exception in the case that the spatial partition
is trying to be created with no rows.