Limit query submission threads
Analysis is the part of the submission and when analysing Presto communicates
with the metastore using the blocking IO. Thus the maximum number of threads
that handle submission remains high (twice the number of CPUs).

Callbacks of the FailedQueryExecution and the ResourceGroupManager are supposed
to be lightweight and low latency. That why the unbounded executor remains to be
used to handle those.