Extend TupleDomainFilter to support filters on nested columns
When applying a filter to a nested column (e.g. an int in an array(int)) one
needs to be aware of the boundaries of the top-level positions. If a single nested
position fails, all positions corresponding to the same top-level position should
fail. The new getPrecedingPositionsToFail and getSucceedingPositionsToFail methods
of TupleDomainFilter allow the caller to know how many positions back and forward
should be marked as failed.

This commit introduces new methods to TupleDomainFilter and updates the callers.
A future commit will introduce implementations of TupleDomainFilter that will apply
to nested columns and return non-zero values from the new methods.