Perform additional determinism analysis for queries with LIMIT clause
We perform determinism analysis when there is a result mismatch. To
do so, we rerun the control query up to 2 more times and check if
the checksum results are the same with the initial run or not.

However, queries with LIMIT may consistently produce the same result
on the control cluster since LIMIT does not guarantee randomness,
even though the query is non-deterministic in nature.

In those cases, we use a best-effort approach to identify if the
query without the LIMIT clause would acutally produce the same
(Deterministic), or a larger number of rows (Non-Deterministic).

The analysis supports INSERT, SELECT, and CREATE_TABLE_AS_SELECT
queries that has a top-level LIMIT clause in its SELECT query part.