Add support for auto-resolving HIVE_TOO_MANY_OPEN_PARTITIONS error
When Presto writes to a bucketed table, each writer is assigned a fixed
set of files (buckets) per partition. If a test cluter is smaller than
than the control cluster, it is possible that each writer is assigned
more files than is allowed, resulting in HIVE_TOO_MANY_OPEN_PARTITIONS.

The failure resolver attempts to estimate the number of expected files
per writer, by looking at the bucket count of the table, and the test
cluster size. Number of partitions generated by the query is assumed
to be 1 for safety purpose, since we currently don't have a good way
to obtain the output partition count for the control query. The
assumption guarantees we're always under-estimating (if not precisely)
the number of files per writer, and therefore we will NOT resolve
a failure with reduced confidence.