Update selective stream readers to avoid mutating input positions
The positions arrays passed to SelectiveStreamReader.read method may be shared
between multiple readers and therefore no reader should modify it. To avoid
unnecessary copies, this change adds a boolean to track whether outputPositions
is a read-only copy of input positions and updates compactValues to make a copy
of outputPositions when needed.