Fix MapBlock.ensureHashTableLoaded
createMapBlockInternal takes positionCount as the logical positionCount
but the hashtables need to be built on the entire underlying MapBlock.
Since offsets.length may be larger than the actual valid number of offsets,
we need to know the number of hashtables from somewhere else. This commit
adds expectedHashTableCount to HashTables and uses it in
MapBlock.ensureHashTableLoaded()