Use TableWriterMergeNode to check recoverability
After we introduced TableWriterMergeNode, the root node of writer
fragment is TableWriterMergeNode, so we should use TableWriterMergeNode
to check recoverability.

To make sure this is properly tested, this change would also enable
only two tests in TestHiveRecoverableGroupedExecution. This test is
known to be flaky when we run too many tests because GC would slow
down query execution and cause timeout. Only enabling one seems
a reasonable compromise.