Add slug for task output
As of now anyone who can build the task output URL can get the data from workers. This change adds a communication slug per task that the fetching task passes to get the output of the upstream task - this means that even if someone can build a task url, then need to know the communication slug to be able to get the data.

The slug is generated per task and then delivered to the workers running the task using the `TaskResource.createOrUpdateTask` function and to the tasks fetching the data using `RemoteSplit`.