Remove synchronized method in FixedSourcePartitionedScheduler
This would resolve a deadlock for the following scenario:

- SqlStageExecution#updateTaskStatus calls FixedSourcePartitionedScheduler#recover
- FixedSourcePartitionedScheduler#schedule calls SqlStageExecution#scheduleSplits

However, in SqlStageExecution, both updateTaskStatus and scheduleSplits
are synchronized, while in FixedSourcePartitionedScheduler, both
schedule and recover are both synchronized. Thus there is a
deadlock.

Instead of making schedule() and recover() synchronized, we could
do recover job at the begining of schedule(). We use a non-blocking
concurrent queue to store the tasks to be recovered, and make
recover() to be the producer and schedule() to be the consumer
of the queue.