Make Polygon WKTs valid in test cases
The standard for polygons requires the loops to be closed -- ie, the
final point should be the same as the first point.  The ESRI library allows
non-closed loops, which it closes implicitly (if the final point is not
the same, add the first point to the end).  Many of our test cases use
this implicit correction, which will fail if we run it under JTS.

This corrects many of those instances.