Fix selective readers to make a copy of input positions
Each reader has a mutable array of output positions. read() populates this
array and returns it via getReadPositions() to the caller who is passing it to
the read() method of the next reader as "input positions". Subsequent calls to
getBlockView() or getBlock() may modify that array by removing positions
filtered out by the other readers.

Readers without filters used to hold on to the input positions array assuming
the array won't change. This assumption is incorrect and may cause a reader to
get into incosistent state. This change makes it so that all readers always
make a copy of input positions in the read() method and never rely on that
array to not change.