Reliable Resource Groups with versioning
Problem:
In case of resource group changes, if we make any leaf resource group to be intermediate resource group by introducing new resource group, all the queries assigned to the old resource group before the change and are queued, will fail when they are being processed. And the reason for that, we fail all queries which try to run on an intermediate resource group. But in this scenario, we should allow those queries to be run on intermediate resource group and not fail.

Solution:
Introducing versioning to Resource Groups. Please find more information on the behavior of the resource group versioning as below.

With every change in the resource group, Resource group manager will create a new version of each of the resource groups.
Queries queued on the old version will move to the new version if the new version resource group is still a leaf node. Otherwise, queued queries stays with old version and continue getting executed as per the old resource group contraints (i.e. hard concurrency)
For every new query to run, resource group manager first checks in the older version if any pending query eligible to run and runs them first before moving to the new version.
For any new resource group, when running query it checks total running query which includes all previous version queries as well. And if that still within the constraint, then only allow the new resource group. While queuing queries, it only checks it's existing limit and queue query accordingly.