Bypass http call if the node is local
If a global config is set and the node-uri matches the current
node, then use the task manager directly for fetching exchange data and
doing task calls. That is bypass the taskResource if these conditions
are satisfied.

In addition, the backoff changed  such that the local mode
never undergoes a backoff. It's a bit hacky but does the job.

Can be enabled (default is disabled) by setting
task.bypass-http-for-local=true