Improve error message for the array_agg function
Currently, array_agg throws a `GENERIC_INTERNAL_ERROR` when an array
is too large. eg: "Cannot allocate slice larger than 2147483639 bytes"
message is reported when the underlying data is stored in Slice.
This was confusing to users.

This commit throws `PrestoException` with
`CELL_TOO_LARGE` error code
when the retained size of a single array exceeds 1G.