Properly handle optimizable filter expressions
Filter expression can be optimized to a constant value (true, false or null) or
to a simpler expression that uses fewer inputs (columns). Make sure to read only
columns that are references from the optimized expression and replace table
scan node with empty values node if filter expression can be optimized to false or null.