Add Alluxio Hive metastore
This change introduces the Alluxio metastore which connects to the Alluxio catalog service. See: https://docs.alluxio.io/os/user/2.1/en/core-services/Catalog.html for more information on the catalog service. In this PR we introduce the bare minimum to configure and use the Alluxio catalog. Reads are the only supported operation at the moment.

Co-authored-by: David Zhu <david@alluxio.com>
Co-authored-by: Zac Blanco <zac@alluxio.com>