Fix like pattern interpretation
like pattern constant folding happens only when the optimization level
is above SERIALIZABLE. However, RowExpressionInterpreter::tryHandleLike
handles like function in a way that it will take the constant folded
like pattern result together with the original arguments. Under such
case, tryHandleLike expects the given constant folded like pattern
should always be a compiled Regex but is not in the above case.

The patch relaxes the restriction on the like pattern to be more
generic. As long as the return type is Regex, tryHandleLike should be
able to handle it.