Add BingTile cast to/from bigint
Externally, tiles are encoded in a string of chars '0' to '3' called a
quadkey.  Internally, Presto encodes a tile in 64 bits, represented by a
BIGINT.  Storing a tile as a bigint is not only more space/cpu efficient
than storing it as a quadkey, but it also avoids the bucket-skew problem
caused by the non-uniform distribution of `hash(quadkey) mod 2^k`.