Improve entropy for bing tile bigint encoding
Java hashes longs by XORing the first 32 bits with the second 32 bits.
Hive assigns buckets based on this hash.  If you have 2^k buckets, you
only keep the lowest k bits of the hash.  Often, k is 9 to 12, and the
previous encoding did not have much entropy in those low bits.  The
resulting first 5 bits were the zoom XORed with bits 5-9 of the x.
If the partition has a constant zoom (very common) and the zoom is less
than 9, several combinations of these bits would be missing, which would
mean empty buckets.

Checking the distribution over 1024 buckets for the old and new hash
function, we get:

Method | min   | mean    | max   | stddev
old    | 12659 | 50313.4 | 89397 | 22831.2
new    | 48344 | 50313.4 | 52626 | 1031.6

The stddev drops by a factor of 20x.