Fix checksum column generation for map keys and map values
Map keys and map values are both arrays. When checksuming an array,
we need to handle the case in which the array element is orderable
but the sorting fails.

The sort of an array would fail in certain cases when the array
elements are array with nulls or row with nulls. As an example,
array_sort(ARRAY[ARRAY[null], ARRAY[1]]) would fail.

This has been properly handled in ArrayColumnValidator, so we should
apply the same logic to the keys checksum and the values checksum
for an map column.