Refactor LocalExecutionPlanner
Add additional "plan" method that takes a OutputFactory as an input.

It is different from the existing "plan" method (that also takes OutputFactory).
The existing version does not create partitioning function, as it is only intended to
use by the LocalQueryRunner that doesn't have to partition the data between stages.
The new version creates a partitioning function based on the PartitioningScheme and passes
it as a parameter of the OutputFactory.

This allows to override OuputOperator behavior without duplicating the code that
creates the partitioning function.