Make sure spark service is created only once
If VM is recycled it is possible to reuse Presto service.
Generally it is not expected for a VM to be reused between different queries.
If (for some strange reason) VM is reused between queries, and the Presto
package or configuration is different for these queries - we prefer to fail
instead of keep creating Presto services. Immediate failure is preferred to
avoid hidden liveness issues (such as memory leaks, etc.)