Remove dependencies from the shaded presto-launcher pom
Shaded artifacts includes a copy of the dependency libraries classes.
The dependency declarations have to be removed from the generated
artifact pom. Otherwise when dependencies are included - classes
are going to be double-included in the project that uses a shaded
jar as a dependency.

Unfortunately there is no way of generating a separate, reduced dependency
pom file for only a "shaded" classified artifact, thus the main artifact
has to be replaced with a shaded one.