Support determinism analysis for simple ORDER BY LIMIT queries
Support determinism analysis for simple queries with top-level ORDER
BY LIMIT clause. Select, Insert, and CreateTableAsSelect are supported,
but only when the query part is a Select query, not a SetOperation(
i.e., Union, Intersect, and Except).

"INSERT INTO ...SELECT ... ORDER BY ... LIMIT ..." is supported.
"INSERT INTO ... SELECT ... UNION ALL ... SELECT ... ORDER BY ...
LIMIT ..." is not supported.

To check for determinism of ORDER BY LIMIT queries, we run the Select
query with limit N + 1, project all the necessary ORDER BY columns,
and check whether there is a tie on the ORDER BY columns for the
n-th row and the (n+1)-th row.