Reduce positions array size in PartitionBuffer
In OptimizedPartitionedOutputOperator, the PagePartioner has
partitionCount number of PartitionBuffer, and each PartitionBuffer
has a positions array to record the positions to be appended to this
buffer. Previously this positions array was initialized to be the
incoming page's positionCount size. This could waste lots of memory
since each partition may only get a small portion of rows. This commit
reduces the initial positions array size from positionCount to
min(positionCount, (positionCount / partitionCount + 1) * 2) and
grow it on the fly.

JMH benchmark shows about 8-10% gain in retained size:
			addPage		optimizedAddPage optimizedAddPage_reducePositions Gain%
bigint	 	 	335,715,874 	 424,471,312 	 	390,497,040 		   8%
ARRAY(BIGINT)	 	327,680,906 	 294,847,046 	 	270,506,390 		   8%
ROW(BIGINT,BIGINT)	315,490,496 	 261,529,566 	 	236,571,362 		   10%