Fix memory tracking for PartitionedOutputOperator
Previously the PartitionedOutputOperator tracks memory after finishing
processing the page. This could cause severe under counting when the
pages are very large and flush every time because flush() resets the
pageBuilders. This commit reports the memory before flush() which is
more accurate. It also changes the counting from getSizeInBytes() to
getRetainedSizeInBytes() because the first was under counting for 20%.