Fix getRetainedSizeInBytes for ColumnarXXX
The ColumnarXXX#getRetainedSizeInBytes() used to add up nullCheckBlock,
nested blocks and offsets's sizes. But nullCheckBlock already covers
the nested blocks and offsets. The nested blocks, e.g. elementsBlock
in ColumnarArray, is just a wrapper of nullCheckBlock's values block.
For a plain block, it is sufficient and close enough to just count
nullCheckBlock's size. For Dictionary or RLE blocks, the retained size
also includes the size of the newly created dictionaryIds and offsets.