Enforce buffer size limits for BlockEncodingBuffer
The byte[] buffers in BlockEncodingBuffer used to be grown by a factor
of 2. This could cause the buffers larger than necessary. Reducing
the expansionFactor has negative impact on performance. Therefore we
want to enforce stricter limit when growing these buffers. This commit
uses incoming blocks' logical sizes as an estimation for how large the
buffers can grow. It shows 30% to 40% memory usage reduction for fixed
width types and 10% reduction for variable width types for the
OptimizedPartitionedOutputOperator.