Optimize metastore communication when writing to partitioned table
Avoid unnecesary metastore calls on Presto workers when writing to
a partitioned table.

Generally it is required to get the partition information on the worker
to figure out the storage format and other partition parameters.

However if the partitions are immutable the append is not supported, thus
the information about the underlying storage format of a partition is not
needed.

However the getPartition call is still needed to check if the partition
exist. As in case when overwrite is not allowed we would like to fail fast.

Given the number of extra metastore calls it results into it feels like
it is not reasonable price to pay for fail fast for this non very common
case.

If the direct write into the destination partition is required (with no
staging directory) the old model should stay in place, as it is the only
way to avoid adding files into the folder of an immutable partition.