Pass the correct value for fileSize in Hive internal split factory
We are passing the split length as `fileSize` to `InternalSplit`
which was used by Parquet reader when trying to find the last few bytes
that contain the Parquet file magic code to verify the file is a Parquet
file. PR #12780 accidentally modified this. It looks like Parquet is the
only reader which is using the filed `fileSize and this reproes only
if `InputFormat.getSplits()` is used which is not the most common.