Support SHOW CREATE FUNCTION
Support syntax, semantic analysis, and execution for SHOW CREATE
FUNCTION. An optional list of parameter types may be specified.

For each matching signature, returns the canonical CREATE FUNCTION
query text that can be used to create the exact same function,
along with the parameter type list of that function signature so
that user can differentiate among the functions.