Remove compression for creating zero row file
Using compression for zero row files for missing buckets
was running into issues with ZSTD compression as it was not
supported by downstream dependencies.
Not using compression while creating zero row files is also better
from efficiency and compression perspective. In my local testing I
observed that for any unused bucket, it was generating 34 bytes with
ZLIB, compared to 27 bytes with no compression.