Remove positionCount check in createRandomLongDecimalsBlock
TestRowBasedSerialization sometimes fails calling
createRandomLongDecimalsBlock with less than 10 positions. We should
allow blocks with less than 10 positions to be created if there are
such needs. This commit removes the check to enforce the positionCount
check, and comments were added to suggest the user use a larger
positinCount when desired nullRate > 0.