Add flatten_geometry_colletions function
GeometryCollections are often problematic for geometry computations.
First, some standard functions perform non-intuitively for them, or may
throw an exception.  Second, they can be arbitrarily nested, so ensuring
that you've flattened a geometry to contain no GeometryCollections is
challenging in SQL.  A common request is to recursively flatten any
encountered GeometryCollections, which is what this function does.

Non-GeometryCollections transform to a singleton array:
`MULTIPOINT(0 0, 1 1)` goes to `[MULTIPOINT(0 0, 1 1)]`.

GeometryCollections are recursively flattened to an array of their
(leaf) constituents:
`GEOMETRYCOLLECTION(POINT(0 0), GEOMETRYCOLLECTION(POINT(1 1)))`
goes to `[POINT(0 0), POINT(1 1)]`.  The order of the array is
implementation-dependent and not guaranteed.