Reduce copies in KdbTree
Currently, KdbTree copies the items for each split, which is
O(listSize * treeHeight).  However, since each parent sorts
before each child, and each child only sees a portion of the
list, the sorting can be done in place.  To prevent these
modifications from affecting the caller's list, there is
an initial single copy, so that all sorts are done on that
copy.