Refactor PrestoSparkRddFactory
This is preparation for Broadcast join implementation

Broadcast join is special, as the broadcasted table must be first collected on the driver,
so it can be broadcasted with a Broadcast variable. This results in more than a single
job per query. This commit prepares the codebase to support more than a single job
per query.

- PrestoSparkRddFactory now creates RDD only for a single fragment, and not for a
  fragment tree. The dependent RDD have to be provided as a parameter. This is needed
  to allow Broadcast sources to be provided.
- SubPlan traversal is now being done in PrestoSparkQueryExecutionFactory. This allows
  us to execute dependency broadcast jobs in the `execute()` method.
- PrestoSparkSplitEnumerator is removed and PrestoSparkPlan is inlined. Splits are lazily
  enumerated in the PrestoSparkRddFactory to avoid enumerating splits that are only necessary
  for the current job being executed. That allows us to save precious memory for the broadcast
  table.