Create a freemarker based Parquet batch ColumnReader generator from template
From the template (ex. Int32FlatBatchReader) generate type specific implementation
class during build time. This reduces the amount of code that we have for the Parquet
reader which is mostly the same across all column reader with some variations.
If we make any changes to one of the reader, it is mostly applicable to other
readers. Now with the template we only need to modify the reader in once place
and it also avoids the causing bugs that are fixed in one reader but not in other
readers.