Support broadcast union
Presto on Spark does not support round robin partitioning.
Round robin partitioning is very expensive in Spark and should be avoided.
For now unions are collected into a single node. Given there's not that many
query shapes that would result in such plan it should be fine as a temporary
solution. However long term we should support multiple table scans in the
fragment to avoid such suboptimal union planning.