Update README to fix VM options
PrestoServer attempts to create directories in "storage.data-directory" for Raptor. This is configured in etc/raptor.properties by default as "/var/data". In OSX, this is not a writable directory, so PrestoServer fails to start. This change adds a VM option that overrides etc/raptor.properties and creates raptor's data directories in $HOME.