Warn when using expensive UNION DISTINCT queries
UNION DISTINCT can be expensive, when there is a lot of data involved.
We are sending a warning when there there are many visible fields, and
some may be expensive to process.  The current criteria are, that there
are greater than 3 visible fields, and that one is a complex type.