Update compiledLambdaMap in RowExpressionCompiler for SQL functions
With the introduction of SQL function, we might find new lambdas while processing SQL functions.
Since we are adding the compiled lambda function to the same class as the expression containing lambda,
we need to update the compiledLambdaMap to make sure it always has all the previous compiled lambdas
and do not try to create the same method again.