Handle mixed-case columns in Elasticsearch
The roundtrip of column names via ColumnMetadata causes the name
to be lower-cased. This results in a mismatch when decoding
the documents produced by Elasticsearch.

This change retains the original column name as an additional
property in the ColumnMetadata for each column.

Cherry-pick of https://github.com/prestosql/presto/commit/0300bd0c0655d38a3fc6d7659bf4463ad925c8b4

Co-authored-by: Martin Traverso mtraverso@gmail.com